+++ lib/screens/home/widgets/widgets.dart
@@ -2,4 +2,3 @@ export 'package:entre_tiempos/screens/home/widgets/app_bar_action.dart';
 export 'package:entre_tiempos/screens/home/widgets/side_bar.dart';
 export 'package:entre_tiempos/screens/home/widgets/info.dart';
 export 'package:entre_tiempos/screens/home/widgets/calendar.dart';
-export 'package:entre_tiempos/screens/home/widgets/tasks.dart';
